app: model: track: Add TrackModel to handle upload and hashing of tracks
Add a TrackModel to handle all things related to tracks. The idea
here is that a track is uploaded to a PHP script and this model handles
hashing the file and then saving that file named as it's hash into the
uploads path defined in the config.

Through this we can avoid duplicates because a file with the same hash
will overwrite another, however the data is identical. Wonder how much
benefit is actually to be seen as encoding artifacts/noise/etc may
change files slightly to make duplicates non-existant?