devops: lamp-node-dev: Add a dev server configuration script
So far we have messed up every single release and needed a few
incremental patches on the production server after deploying to get
things all the way working. This patch attempts to add a development
subdomain and all necessary support items to trackdrop.io so that we can
first deploy to dev.trackdrop.io (or perhaps some obfuscated url) before
moving things to trackdrop.io.

This should deploy the same with the only difference being this assumes
the main server has already been deployed and only adds to the nginx
config rather than building it from the ground up.