core: router: Clean up and simplify API routing
We don't need to be so specific with API routing, let's change it so
that API routes look like:

	http://trackdrop.io/api/v1/tracks

So that APIs are only partitioned by major version, which should signify
a compatibility breakage when it moves. This also means that API routes
are stored as:

	config/routes/api/v1.php

and API controllers are stored as:

	app/controller/api/v1/track.controller.php

So everything is greatly simplified.