trackdrop-client: Refactor main content into separate UserTracks component
Refactor the main content, which currently consisted of a single
Container component with the upload box and track list in it, into its
own new 'UserTrack' component. Additionally, move the initial track
loading logic into same component.

This is done in preparation of introducing some routes and making use of
react router.