trackdrop-client: Introduce Routes for Login, Register, Tracks
As more functionality is introduced we cannot maintain a "true" single
page application, so we must introduce React Router and move the current
"Tracks" homepage into a route at "/tracks". While we are at it,
introduce routes for two new pages that provide functionality to support
login, "/login" and "/register".

Introduce simple placeholder components for Login and Register
which have basic forms but no action yet.