controller: Ensure correct headers are being set with json response
It seems that the status code was not being set at all and all responses
in JSON were receiving 200 OK regardless of the actual status. By
setting the status code of the last call to header in the controller
output code, we can get the right status to be returned.