trackdrop-client: Introduce stateful user login/out and register
This is a massive commit to introduce all required actions, reducers,
helpers, routes, and state tracking to allow for users to register
accounts, login, and logout.

The '/tracks' page and uploading are now protected by a session and
require that a user login.

Additionally, this commit introduces the plumbing for redux and the
first two redux actions/reducers for login/logout.