trackdrop-client: tracklist: Handle empty initial tracklist
Sometimes the tracklist will be empty, and even more commonly now
because users can create accounts and may not have any tracks available
immediately, so we must handle the case where none are available by
checking the number of tracks we've pulled and displaying a message if
nothing is there.

NOTE: Message must still be styled.