client: trackbox: Update delete API call to consider logged_in status
We must ensure a user is logged in before letting them delete a track,
so update the API call to use an auth call from the new trackService.

In order to fully understand logged in user data, we must introduce
redux for the trackbox, so connect it, which also changes how the class
is exported so fix the includes in other files as well.