model: track: Do not fail if move_uploaded_file fails due to existing file
When uploading a file the move_uploaded_file function can fail if there
is an existing file with the same name present in the file storage area.
This is not a true failure case, and identical files are automatically
de-duplicated through the hash name of the file, so do not fail.